+++ app/src/main/java/com/example/btppilot/data/dto/response/tasks/TasksByProjectDtoItem.kt
@@ -11,7 +11,7 @@ data class TasksByProjectDtoItem(
     @Json(name = "description")
     val description: String,
     @Json(name = "estimationHours")
-    val estimationHours: Int,
+    val estimationHours: Double,
     @Json(name = "plannedStartDate")
     val plannedStartDate: String,
     @Json(name = "plannedEndDate")
